pyproject.toml
[tool.poetry]
name = "xtrades"
version = "0.1.0"
description = "A Python daemon that monitors IBKR trades and posts updates to X"
authors = ["Mohammad Dashti <mdashti@gmail.com>"]
readme = "README.md"
packages = [
    { include = "*", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
ib-insync = "^0.9.86"
tweepy = "^4.15.0"
sqlalchemy = "^2.0.37"
schedule = "^1.2.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^4.1.0"
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.9.0"
pytest-asyncio = "^0.23.5"
pyright = "^1.1.392.post0"
ruff = "^0.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"  # Removed --cov flag
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow running"
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
